fix: add generateStaticParams for static export compatibility
- Split page.tsx into server component with generateStaticParams
- Move client code to TaskDetailClient.tsx
- Required for Next.js output: export with dynamic routes
